rename to visualization-frame.component
