+++ core/new-gui/src/app/workspace/component/result-panel/visualization-frame/visualization-frame.component.html
